app: promote the desktop app to be the Electron demo
