--- apps/desktop/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "runanywhere-ai-desktop",
-  "productName": "RunAnywhere AI",
-  "version": "0.1.0",
-  "description": "Private, on-device AI for Windows — chat, reasoning, your own knowledge base, voice, and vision. Nothing leaves your machine.",
-  "author": "RunAnywhere",
-  "license": "SEE LICENSE IN ../../LICENSE",
-  "private": true,
-  "main": "main.js",
-  "scripts": {
-    "start": "electron .",
-    "start:gpu": "electron . --gpu",
-    "icon": "electron assets/make-icon.js"
-  }
-}
